+++ Demo/Sources/OnDeviceVLMSDKDemoApp.swift
@@ -1,6 +1,6 @@
 //
-//  GlimpseDemoApp.swift
-//  GlimpseDemo
+//  OnDeviceVLMSDKDemoApp.swift
+//  OnDeviceVLMSDKDemo
 //
 //  Created by Nazar Kozak on 12.06.2026.
 //
@@ -8,7 +8,7 @@
 import SwiftUI
 
 @main
-struct GlimpseDemoApp: App {
+struct OnDeviceVLMSDKDemoApp: App {
     var body: some Scene {
         WindowGroup {
             ContentView()